refactor: components/donate/Result.vue -> components/app/Donate/RdDonateResult.vue
